Update the ResponseWrapData function to return a wrapping.ResponseWrapInfo object
